pyproject.toml
[tool.poetry]
name = "django-serializers"
version = "0.1.0"
description = ""
authors = ["hsanfilippo <henrique.sanfilippo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.3.2"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"
django = "^5.1"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
